Back/django-game/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.5
django-allauth==0.45.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
# djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
flex==6.14.1
hangul-utils==0.2
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonpointer==2.1
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
pycparser==2.20
PyJWT==2.1.0
# mysqlclient==2.0.3
PyMySQL==1.0.2
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rfc3987==1.3.8
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
strict-rfc3339==0.7
uritemplate==3.0.1
urllib3==1.26.6
validate-email==1.3
